+++ src/Buildalyzer/AnalyzerManager.cs
@@ -1,12 +1,14 @@
-﻿using System;
+﻿extern alias StructuredLogger;
+
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Buildalyzer.Logging;
 using Microsoft.Build.Construction;
-using Microsoft.Build.Logging.StructuredLogger;
 using Microsoft.Extensions.Logging;
+using StructuredLogger::Microsoft.Build.Logging.StructuredLogger;
 
 namespace Buildalyzer
 {