+++ src/Buildalyzer/Logging/PropertiesAndItems.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections;
+
+namespace Buildalyzer.Logging
+{
+    internal class PropertiesAndItems
+    {
+        public IEnumerable Properties { get; set; }
+        public IEnumerable Items { get; set; }
+    }
+}